+++ src/dotnet/MediatorPlugin/Services/Create/IHandlrCreator.cs
@@ -1,7 +1,7 @@
 ﻿using JetBrains.ReSharper.Psi.CSharp.Tree;
 using JetBrains.ReSharper.Psi.Tree;
 
-namespace ReSharper.MediatorPlugin.Services
+namespace ReSharper.MediatorPlugin.Services.Create
 {
     public interface IHandlrCreator
     {