+++ src/dotnet/MediatorPlugin/Services/Navigation/INavigationOptionsFactory.cs
@@ -1,0 +1,11 @@
+using JetBrains.ReSharper.Feature.Services.Navigation;
+
+namespace ReSharper.MediatorPlugin.Services.Navigation;
+
+internal interface INavigationOptionsFactory
+{
+    NavigationOptions Get
+    (
+        string title
+    );
+}