+++ CHANGELOG.md
@@ -3,6 +3,7 @@ All notable changes to this project will be documented in this file.
 
 ## 25.05.XXXX
 - Support for 2025.2
+- Support for multiple handlers implementing one request
 - MediatR is going commercial so this extension supports another MediatR alternative: "https://github.com/martinothamar/Mediator"
 
 ## 25.03.16XX